Refactor step 3: Separate vrfy and vrfyCli
